+++ Firmware/Src/gpio.c
@@ -191,15 +191,46 @@ void SetGPIO12toStepDir() {
   GPIO_InitStruct.Pull = GPIO_NOPULL;
   HAL_GPIO_Init(GPIO_2_GPIO_Port, &GPIO_InitStruct);
 
+  //TODO: Hardcoded EXTI line not portable. Get mapping out of Cubemx by setting EXTI default
   HAL_NVIC_SetPriority(EXTI0_IRQn, 0, 0);
   HAL_NVIC_EnableIRQ(EXTI0_IRQn);
 }
+
+//TODO: Enable index on only one channel
+void SetupENCIndexGPIO(){
+  GPIO_InitTypeDef GPIO_InitStruct;
+
+  /*Configure GPIO pins : PAPin PAPin */
+  GPIO_InitStruct.Pin = M0_ENC_Z_Pin;
+  GPIO_InitStruct.Mode = GPIO_MODE_IT_RISING;
+  GPIO_InitStruct.Pull = GPIO_NOPULL;
+  HAL_GPIO_Init(M0_ENC_Z_GPIO_Port, &GPIO_InitStruct);
+
+  //TODO: Hardcoded EXTI line not portable. Get mapping out of Cubemx by setting EXTI default
+  HAL_NVIC_SetPriority(EXTI15_10_IRQn, 0, 0);
+  HAL_NVIC_EnableIRQ(EXTI15_10_IRQn);
+
+  /*Configure GPIO pins : PBPin PBPin */
+  GPIO_InitStruct.Pin = M1_ENC_Z_Pin;
+  GPIO_InitStruct.Mode = GPIO_MODE_IT_RISING;
+  GPIO_InitStruct.Pull = GPIO_NOPULL;
+  HAL_GPIO_Init(M1_ENC_Z_GPIO_Port, &GPIO_InitStruct);
+
+  //TODO: Hardcoded EXTI line not portable. Get mapping out of Cubemx by setting EXTI default
+  HAL_NVIC_SetPriority(EXTI3_IRQn, 0, 0);
+  HAL_NVIC_EnableIRQ(EXTI3_IRQn);
+}
+
 
 //Dispatch processing of external interrupts based on source
 void HAL_GPIO_EXTI_Callback(uint16_t GPIO_Pin) {
   //Step signals for M0 and M1
   if (GPIO_Pin & GPIO_1_Pin || GPIO_Pin & GPIO_3_Pin) {
     step_cb(GPIO_Pin);
+  } else if(GPIO_Pin & M0_ENC_Z_Pin){
+    enc_index_cb(GPIO_Pin, 0);
+  } else if(GPIO_Pin & M1_ENC_Z_Pin){
+    enc_index_cb(GPIO_Pin, 1);
   }
 }